[Feat] 모임후기 삭제 전 확인 메시지 추가
모임후기 삭제 버튼 클릭시 바로 삭제되지 않고 한번 더 체크하는 모달 추가

+++ src/components/meeting/review/ReviewDetail.jsx
@@ -18,11 +18,27 @@ const ReviewDetail = () => {
       .catch((err) => console.log(err));
   }, []);
 
-  const onDeleteHandler = () => {
-    apis
-      .deleteMeetingReview(params.id)
-      .then((res) => navigate('/review'))
-      .catch((err) => console.log(err));
+  const onDeleteHandler = (e) => {
+    e.preventDefault();
+    swal('정말로 삭제하시겠습니까?', {
+      buttons: {
+        cancel: '취소',
+        삭제: true,
+      },
+    }).then((value) => {
+      switch (value) {
+        case '삭제':
+          apis
+            .deleteMeetingReview(params.id)
+            .then((res) => navigate('/review'))
+            .catch((err) => console.log(err));
+
+          break;
+
+        default:
+          break;
+      }
+    });
   };
 
   return (
